// src/app/models/personal-data.ts
export class PersonalData {
   id?: number;
   creationDate?: Date;
   modificationDate?: Date;
   isDeleted?: true;
   nationality?: string;
   dni: string;
   address: string;
   countryCode: string;
   regionCode: string;
   postalCode: string;
   phone: string;
   name?: string;
   surname?: string;
}